Update README.md
I started this project more than a year ago before the steam version was released. Upon news of the steam version being released I lost interest on this version and the project collected some dust for a while. While trying to keep my typescript skills up to date I got back into this project in the beginning of 2019. Once I got it to a functional place for the original board I was mostly losing interest and was planning to move on to a different board game in 2020. 

Somewhere along the way in 2019 more people showed up and started contributing! We need to include more updates to our `README.md` to call out that the prelude expansion and the venus expansion are now supported but I figured we should start with a contributors section since there are more people giving their time to this project for one reason or another. Your contributions have motivated me to keep with this project wherever it may go.