Fix english typos (#732)
* "loose" -> "lose".

This change affects only visible text.

* "Scraping" -> "Scrapping".

"Scrapping" is the language used in the printed game, on both the rules and the cards.
It is probably a typo, but this makes the usage consistent. This change affects both
symbols and visible text.

* STANDART -> STANDARD.

This change affects only symbols.

* STANDART -> STANDARD.

This change affects only symbols.