Update LocalHeatTrapping.ts
You cannot use megaCredits as heat. To play this card you MUST have 5 heat.
The tricky part is if the player can use heat as megaCredits then they not only need 5 heat but ALSO heat + megaCredits >= 5 + cardCost
I fixed the return value and removed the spurious code that deducted MC if there was not enough heat.