Clarify PlayerTags & PlayerInfo.
First: activePlayer is not a great term: the active player is the one
taking the current action, but here it means "me, the player with the
brower."

Second: Anyone who is NOT "me, the active player" is of type
PublicPlayer, not PlayerView, so using the right types will clear things
up.

There's a minor change for a component key to line it up with the
others, too.