Rename CardMetadata -> ICardMetadata.
As I start to detach the card metadata interface, I'm going to take
advantage of the (inconsistently applied) standard in this project to
start interfaces with I. As I've been prototyping a solution, not
knowing what's a class and what's an interface has been a significant
point of confusion.

Note that some of the changes include removing CardMetadata for some
cards and renaming the interface in others. For some reason I don't
yet understand, some of those parameters need the declaration.
So I'm not sure if I'm causing a breakable issue by doing this.