Rewrite GamesOverview
Using a child component which simplifies quite a lot, really.
Also not sending all the requests at once (what a thing!)
Add strong typing everywhere I can.

This whole thing would be faster if the API accepted a large number of
IDs at a time. That can be next time.