Improve game-end state management.
1. Ensure that the game is saved at game.end.
2. When deserializing a game, don't do anything if the game is at end
phase.

What was happening was that the server resets and then the players
somehow get to reload their game and also find that they're in their
final action phase again.

Testing was hard because there are still some things that aren't async.
I manually tested these, including verifying that the game wasn't saving
the end phase.