Inline loadGame.
Now that loadGame is only called in one spot, it can be inlined.

Also, it is now apparent from the prior version that the part of
loadGame which checked the cache is no longer necessary, so that was
removed.

Also added documentation to clarify why getGame's map checks are the way
they are.