Speed up ID initialization.
This creates a new table to track game IDs. It speeds up load time immensely. Loading up a copy of the production database running on my workstation:

old: 36 seconds
new: ~ 1s.

Keep in mind that the same operation on Heroku takes over 10 minutes, so this change
is even more dramatic. (10 minutes to est. 20 seconds?)

This isn't ready for general use, though. It's missing a very seamless way for new people to use it (Not to mention there's no SQLite implementation.)

But as is, this seems the right direction.