Start moving to a multi-corporation environment.
This introduces `getCorporation` which, despite the eventual Merger
prelude is turning out to be a pretty useful slice of code.

`corporations` as a getter is a starting point for eventually moving to
a multiple corporation setup. It's possible this might not stay the
same name (might even move away from corporations to all played
cards. But it reduces ther reliance on a single `corporationCard`