Add initial Countability to Behavior. (#4963)
* Add initial Countability to Behavior.

This initial change will only impact three cards that have a benefit based
on tags in play. Most cards that count tags for behavior increase
production and that's a little more complicated. Better to show this
simple case first, applied to just a couple of cards.

* Make add-resources-to-self behavior Countable.

* Add the ability to specify counting units as a whole.

This addresses a whole slew of behaviors like "Gain 1 M€ per city tile".
I'm kind of excited about it.

* Fix test.

* Fix engineer. Test is coming.

* Code review comments.