Update ui.json
typo
adding translation for all missing mielstones and awards

There is an inconsistency: After all awards there is not a full stop behind the description except for the moon awards. I took the full stop with the moon to not block that part of the translation. The full stop must be removed, but then it need to remove in translation files as well.