Add RunNTimes.
This is a Deferred Action helper for running the same action
several times, producing input results each time.

Very often a series of deferred actions don't depend on a prior one, and
those don't need RunNTimes -- just defer the same action in a loop.

But there are cases, like, selecting spaces, where you don't want to
make the same space eligible in a series of selections.