Move all milestone and award images to one directory. (#7301)
This is going to simplify a few things. First as you see, it simplifies
most of the declarations. After this, it's going to simplify slightly
resizing all the images so they have generally the same shape. All the
MA images if you look at them have extra content that isn't shown. That
will all be trimmed.

There are still more MAs that we can add, and this will just simplify
all of that.